feat: Update NavLinkStyle to use the name prop instead of isTicket ([FEAT] 홈 & 검색 페이지 구현 #3)
The NavLinkStyle component in Navigation.tsx has been updated to use the name prop instead of the isTicket prop. This change allows for more flexibility in styling the navigation links based on the name of the SVG icon.

Refactor the code to use the updated prop and remove the isTicket prop.

Closes #<issue_number>